# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para operaciones frecuentes de docentes y estudiantes en EduGo",
        "title": "EduGo API Mobile",
        "termsOfService": "http://edugo.com/terms/",
        "contact": {
            "name": "Equipo EduGo",
            "email": "soporte@edugo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/materials/{id}/progress": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra el progreso del estudiante en la lectura de un material",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Actualizar progreso de lectura",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del material",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Progreso del estudiante",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateProgressRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/materials/{id}/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene progreso y puntajes de todos los estudiantes para un material (solo docentes)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Obtener estadísticas del material",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del material",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MaterialStatsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "No eres docente de este material",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/materials/{id}/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene el resumen generado por IA desde MongoDB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Obtener resumen del material",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del material",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MaterialSummaryResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Resumen no disponible aún",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                },
                "message": {
                    "type": "string",
                    "example": "Field 'title' is required"
                }
            }
        },
        "GlossaryTerm": {
            "type": "object",
            "properties": {
                "definition": {
                    "type": "string",
                    "example": "Programa que traduce código fuente a código máquina"
                },
                "order": {
                    "type": "integer",
                    "example": 1
                },
                "term": {
                    "type": "string",
                    "example": "Compilador"
                }
            }
        },
...
        "MaterialStatsResponse": {
            "type": "object",
            "properties": {
                "average_progress": {
                    "type": "number",
                    "example": 65.5
                },
                "average_score": {
                    "type": "number",
                    "example": 82.3
                },
                "total_completed": {
                    "type": "integer",
                    "example": 45
                },
                "total_views": {
                    "type": "integer",
                    "example": 150
                }
            }
        },
...
        "MaterialSummaryResponse": {
            "type": "object",
            "properties": {
                "glossary": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/GlossaryTerm"
                    }
                },
                "processing_metadata": {
                    "$ref": "#/definitions/ProcessingMetadata"
                },
                "reflection_questions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "¿Cómo aplicarías este concepto?",
                        "¿Qué diferencias encuentras?"
                    ]
                },
                "sections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/SummarySection"
                    }
                }
            }
        },
        "ProcessingMetadata": {
            "type": "object",
            "properties": {
                "language": {
                    "type": "string",
                    "example": "es"
                },
                "model": {
                    "type": "string",
                    "example": "gpt-4"
                },
                "nlp_provider": {
                    "type": "string",
                    "example": "openai"
                },
                "processing_time_seconds": {
                    "type": "integer",
                    "example": 45
                },
                "prompt_version": {
                    "type": "string",
                    "example": "v1.2"
                },
                "tokens_used": {
                    "type": "integer",
                    "example": 3500
                }
            }
        },
...
        "SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "SummarySection": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Pascal fue desarrollado..."
                },
                "difficulty": {
                    "type": "string",
                    "enum": [
                        "basic",
                        "medium",
                        "advanced"
                    ],
                    "example": "basic"
                },
                "estimated_time_minutes": {
                    "type": "integer",
                    "example": 5
                },
                "order": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Contexto Histórico"
                }
            }
        },
        "UpdateProgressRequest": {
            "type": "object",
            "required": [
                "progress"
            ],
            "properties": {
                "progress": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0,
                    "example": 75.5
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token con formato: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}